Moved DashBoard's Widgets to new folders
