+++ src/pages/Dashboard/Earned/PngToggle/index.tsx
